Inlined file credentials class.
