+++ python/gee-px-ls.py
@@ -2,10 +2,28 @@ import ee
 import numpy
 import os
 from datetime import datetime
-from filecredentials import FileCredentials
+from oauth2client.client import OAuth2Credentials
+
+
+class FileCredentials(OAuth2Credentials):
+  def __init__(self, access_token_file):
+    access_token = self._read_access_token(access_token_file)
+    super(FileCredentials, self).__init__(access_token, None, None, None, None, None, None)
+    self.access_token_file = access_token_file
+
+  def _refresh(self, http):
+    self.access_token = self._read_access_token(self.access_token_file)
+
+  def _read_access_token(self, access_token_file):
+    with file(access_token_file) as f:
+      return f.read()
+
+  def __str__(self):
+    return 'FileCredentials(' + self.access_token_file + ')'
+
 
 # earth engine API init
-access_token_file = os.getenv("HOME") + '/.google_access_token'
+access_token_file = os.getenv("HOME") + '/.google-access-token'
 if os.path.exists(access_token_file):
   ee.Initialize(FileCredentials(access_token_file))
 else:
@@ -109,3 +127,20 @@ try:
 except ee.ee_exception.EEException:
   # no data available
   serie = []
+
+
+class FileCredentials(OAuth2Credentials):
+  def __init__(self, access_token_file):
+    access_token = self._read_access_token(access_token_file)
+    super(FileCredentials, self).__init__(access_token, None, None, None, None, None, None)
+    self.access_token_file = access_token_file
+
+  def _refresh(self, http):
+    self.access_token = self._read_access_token(self.access_token_file)
+
+  def _read_access_token(self, access_token_file):
+    with file(access_token_file) as f:
+      return f.read()
+
+  def __str__(self):
+    return 'FileCredentials(' + self.access_token_file + ')'
